chore(docs): Documenting validation apis
