+++ src/Uno.Extensions.Validation.Fluent/GlobalUsings.cs
@@ -1,9 +1,10 @@
-﻿global using System;
-global using System.Collections.Generic;
+﻿global using System.Collections.Generic;
 global using System.ComponentModel.DataAnnotations;
+global using System.Linq;
 global using System.Threading;
 global using System.Threading.Tasks;
-global using ValidationResult = System.ComponentModel.DataAnnotations.ValidationResult;
-global using Microsoft.Extensions.Hosting;
 global using Microsoft.Extensions.DependencyInjection;
-global using System.Linq;
+global using Uno.Extensions.Validation;
+global using Uno.Extensions.Validation.Fluent;
+
+global using ValidationResult = System.ComponentModel.DataAnnotations.ValidationResult;
