+++ src/Uno.Extensions.Validation.Fluent/ValidationBuilderExtensions.cs
@@ -1,7 +1,17 @@
-﻿namespace Uno.Extensions.Validation;
+﻿namespace Uno.Extensions;
 
+/// <summary>
+/// Extensions for <see cref="IValidationBuilder"/>
+/// </summary>
 public static class ValidationBuilderExtensions
 {
+	/// <summary>
+	/// Registers a fluent validator for the specified type.
+	/// </summary>
+	/// <typeparam name="TEntity">The type to be validated</typeparam>
+	/// <typeparam name="TValidator">The validator to register</typeparam>
+	/// <param name="builder">The validation builder</param>
+	/// <returns>The updated validation builder</returns>
 	public static IValidationBuilder Validator<TEntity, TValidator>(
 		this IValidationBuilder builder)
 		where TValidator : class, FluentValidation.IValidator<TEntity>
