+++ src/Uno.Extensions.Validation/GlobalUsings.cs
@@ -1,10 +1,13 @@
 ﻿global using System;
 global using System.Collections.Generic;
-global using System.Text;
+global using System.ComponentModel;
 global using System.ComponentModel.DataAnnotations;
+global using System.Linq;
 global using System.Threading;
 global using System.Threading.Tasks;
-global using System.Linq;
-global using Microsoft.Extensions.Hosting;
 global using Microsoft.Extensions.DependencyInjection;
-global using System.ComponentModel;
+global using Microsoft.Extensions.Configuration;
+global using Microsoft.Extensions.Hosting;
+global using Uno.Extensions.Validation;
+// Avoids naming clash with System.ComponentModel.DataAnnotations.Validator
+global using Validator = Uno.Extensions.Validation.Validator;
