+++ src/Uno.Extensions.Validation/HostBuilderExtensions.cs
@@ -1,8 +1,19 @@
-﻿namespace Uno.Extensions.Validation;
+﻿namespace Uno.Extensions;
 
+/// <summary>
+/// Extensions for <see cref="IHostBuilder"/> to add validation.
+/// </summary>
 public static class HostBuilderExtensions
 {
 	private static bool _isRegistered;
+
+	/// <summary>
+	/// Registers the validation services.
+	/// </summary>
+	/// <param name="hostBuilder">The host builder to register validation services with</param>
+	/// <param name="configureDelegate">Callback for configuring host builder</param>
+	/// <param name="configure">Callback for configuring additional validation services (eg Fluent)</param>
+	/// <returns></returns>
 	public static IHostBuilder UseValidation(
 		this IHostBuilder hostBuilder,
 		Action<HostBuilderContext, IServiceCollection>? configureDelegate = default,
