+++ src/Uno.Extensions.Validation/IValidationBuilder.cs
@@ -1,5 +1,8 @@
 ﻿namespace Uno.Extensions.Validation;
 
+/// <summary>
+/// Interface for configuring additional validation services
+/// </summary>
 public interface IValidationBuilder : IHostBuilder
 {
 }
